Final Phan cua Minh

+++ PH1/PH1/Source/Form_DeleteUser.cs
@@ -20,12 +20,19 @@ namespace PH1
 
         private void btn_XoaUser_Click(object sender, EventArgs e)
         {
+            username = txt_UserName.Text.Trim().ToString().ToUpper();
+            if (Functions.isUserValid(username) == 0)
+            {
+                MessageBox.Show("User khong ton tai!", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                return;
+            }
             string sql = "alter session set  \"_ORACLE_SCRIPT\" = true";
             Functions.RunSQL(sql);
             username = txt_UserName.Text.Trim().ToString();
 
+            if (chk_CASCADE.Checked == true) 
             sql = "DROP USER " + username + " CASCADE";
-
+            else
             sql = "DROP USER " + username;
 
             if (Functions.RunSQLwithResult(sql) == 1)
